Unexported Name and added zerolog
I realized exporting Name was a bad idea and it would be better to add a Name() method to the Packet interface. 

Added the first dependency, zerolog. Unsure if this is a good choice right now but instead of rolling my own (worse) logger I opted for adding a dependency. Having nice structured logs should making debugging things easier down the road.